Added transform component
Game objects will use components, or maybe it'll be an ecs? Not sure yet, but components get used for sure, so I added a transform component to easily calculate an individual objects model matrix. When GameObject is a thing, it will store one of these as well as model/texture related stuff. Also, added some helpers to mesh to load a cube and quad just for testing. And in texture, if the file couldn't be opened, we fallback to creating solid-colour texture instead.