Resolution sampling, skybox & misc
Main 2 additions, I now have access to frame and render buffers. I use a render buffer for the depth and stencil (not actually making use of stencil yet but still). I also create a texture for the color attachment so we can bind it and draw it to a full-screen quad in the blit pass. Right now, the frame buffers only purpose is to allow me to render at a lower resolution, but having access to post-processing is nice too. I also added skyboxes via a cubemap system. Nothing crazy there really. In terms of misc - added some basic directional gouraud shading and added some typedefs to most the basic gl types, since all the types being u32 made the code harder to parse in my opinion.