hashmap
Added a basic generic hashmap struct. For now we just use it for vertex deduplication, which has sped the function up ~4x. To make it work vertex needed a hash and eq function which are clearly a bit tacked on so those will likely be reworked just to look a bit nicer in the future