spritebatching + some render abstraction + misc
1 - spritebatching, we now have a way to easily draw sections of a texture to the screen as quads. This can be used for 2d and UI and I'm gonna look into using it for text as well.

2 - renderer, the renderer file isn't much for now, it really just abstracts the framebuffer business, but gonna try and put more over there soon because main is real cramped.

misc - Added functions to get window width and height so I can use them to call viewport in renderer (worth noting the window still doesn't support resizing. If that gets added I'll have to update quite a bit). Texture is now a struct that also contains width and height as well as the u32 id, this was important for spritebatch but it meant a few function signatures had to get updated. Finally started on a math module. For now its just a rectangle struct and vector2 and 3 structs. Vertex now uses these instead of the vecs from linmath. This is fine as I dont really perform math on the vertex info, but with structs like camera I'll likely keep linmath vecs.