Terrain gen + GameObject update
Main change is that I added some basic terrain generation using fastnoise. It doesn't support much right now (e.g. no texture blending or anything) but it is working. Also, GameObject now stores function pointers that are used in the game loop, one is for input polling and the other is general per-frame update. The only object to use it right now is the 'player' and its update method simply moves it forwards based on its local forward direction, and its update function allows it to reposition itself. Its supposed to be a bit of a flight controller but it isn't really working yet. Also, instead of the spectator style cam, I added a new chase-style camera that can be attached to a game object, and will simply position itself based on the position of the object.