Controller fix
'Plane' controls now work alot better. To fix I had to have the gameobject store the full matrix instead of constructing it when needed, should be less work im doing now in  exchange for game object having to store slightly more and rotating objects is now a bit more verbose with needing to call linmath functions. This probably still isn't great for flight and I'll need to use quaternions in the future but nice for now